requirements.txt
Package                        Version
------------------------------ --------
attrs                          23.1.0
certifi                        2023.5.7
charset-normalizer             3.2.0
exceptiongroup                 1.1.2
Faker                          19.1.0
h11                            0.14.0
idna                           3.4
outcome                        1.2.0
pip                            23.2
PySocks                        1.7.1
python-dateutil                2.8.2
python-dotenv                  1.0.0
PyYAML                         6.0.1
requests                       2.31.0
robotframework                 6.1
robotframework-faker           5.0.0
robotframework-pythonlibcore   4.2.0
robotframework-requests        0.9.5
robotframework-seleniumlibrary 6.1.0
selenium                       4.9.0
setuptools                     65.5.0
six                            1.16.0
sniffio                        1.3.0
sortedcontainers               2.4.0
trio                           0.22.2
trio-websocket                 0.10.3
urllib3                        1.26.16
wrapt                          1.15.0
wsproto                        1.2.0
